requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.4
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
